Load plugins early for improved flexibility (fixes chfoo/wpull#383)
Plugins are attached to existing HookableMixin instances as before. In addition, they are also kept in an attribute of the HookableMixin class to allow attaching when an instance is created after the plugin has been loaded.

This commit restores the plugin loading behaviour of version 1.2.3. It therefore breaks backward compatibility in some cases; plugins written for versions 2.0.x may need to be adapted.